* code review fixes: Use code language 1.5, tiny refactoring
